Update Unit Tests & Formatting
